refactor(svelte): viewport stuff
